move JAX fatigue script to mjx folder
